Prune arglist before saving sessions
:mksession saves opened buffers and files listed in the arglist.

Assuming you would do:

    $ vim foo bar
    :SS session
    :x
    $ vim
    :SL session
    :bwipeout bar
    :x

In that case the only file left in the session should be 'foo'. But
that's not the case, since :mksession! also saves the argument list,
which still consists of 'foo' and 'bar'. Next time you load the session,
both buffers would be opened again.

But since the argument list is seldomly used, or at least not with the
arguments that were given to Vim, it will be pruned from now on.

References #106.