Set buftype=nofile
The buftype option determines, if Vim considers a buffer to be an actual
file or not and since the startify buffer does not really relate to any
actual file, it makes sense to also set the 'buftype' option to be
"nofile".

In addition, this fixes the bug that vim-airline considers the current
buffer to be '[noperm]', since it is

`&readonly && !filereadable(bufname('%'))`

Also change the corresponding Vader test.