Place custom header on first line
Startify appends to the current buffer using `'$'` causing the first line to be left blank. This can be fixed by having the first `append()` start at `0`. However, this causes the default header to have no gap line at the top and also causes custom headers to have an extra gap line after them. To fix this, the changes are made only if there is a custom header, otherwise the behavior is the same as before.